pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "fastapi>=0.122.0",
    "llama-index>=0.14.8",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-vector-stores-milvus>=0.9.4",
    "openai>=2.8.1",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pymilvus>=2.6.4",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
